requirements.txt
beautifulsoup4==4.12.2
Flask==3.0.0
requests==2.31.0
pandas==2.1.3
matplotlib==3.8.2
python-dotenv==1.0.0
lxml==4.9.3
